requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.2
django-crispy-forms==2.0
django-environ==0.10.0
psycopg2==2.9.6
sqlparse==0.4.4
tzdata==2023.3
